Refactor to sanity checker module
